# packages/contracts/openapi/zhili.openapi.yaml
openapi: 3.1.0
info:
  title: Zhili Logistics AI API
  version: 0.1.0
  summary: Contract-first API for 智立科技物流AI系统
  description: >-
    Public and first-party contract. Mutating domain commands require an
    Idempotency-Key; commands against an existing aggregate also require
    If-Match. Permissions are documented with x-permission and successful
    commands emit the x-audit-event event.
  license:
    name: AGPL-3.0-only
    identifier: AGPL-3.0-only
servers:
  - url: /api/v1
security:
  - cookieAuth: []
...
paths:
...
  /customers:
    post:
      tags: [Customers]
      operationId: createCustomer
      summary: Create a tenant-scoped customer
      x-feature-id: MDM-03
      x-permission: customer.write
      x-audit-event: customer.created
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerRequest'
      responses:
        '201':
          description: Customer created in current tenant.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /orders:
    get:
      tags: [Orders]
      operationId: listOrders
      summary: List order projections using a stable cursor
      x-feature-id: ORD-01
      x-permission: order.read
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Stable order page ordered by updatedAt and id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Orders]
      operationId: createOrderDraft
      summary: Create an order draft
      x-feature-id: ORD-01
      x-permission: order.create
      x-audit-event: order.draft.created
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderDraftRequest'
      responses:
        '201':
          description: Draft created.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationFailed'
...
  /quotes:
    get:
      tags: [Quotes]
      operationId: listQuotes
      summary: List quote projections using a stable cursor
      x-feature-id: RATE-06
      x-permission: quote.read
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Stable quote page ordered by createdAt and id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteListResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Quotes]
      operationId: createQuote
      summary: Calculate a versioned multi-channel quote
      x-feature-id: RATE-06
      x-permission: quote.create
      x-audit-event: quote.created
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuoteRequest'
      responses:
        '201':
          description: Quote with immutable calculation input snapshot.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationFailed'
...
  /quotes/{quoteId}:accept:
    post:
      tags: [Quotes]
      operationId: acceptQuote
      summary: Accept one quote option
      x-feature-id: RATE-07
      x-permission: quote.accept
      x-audit-event: quote.accepted
      parameters:
        - $ref: '#/components/parameters/QuoteId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcceptQuoteRequest'
      responses:
        '200':
          description: Accepted quote version.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteResponse'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '410':
          $ref: '#/components/responses/Expired'
  /orders/{orderId}:link-accepted-quote:
    post:
      tags: [Orders]
      operationId: linkAcceptedQuoteToOrder
      summary: Link an accepted quote option to an order and create its draft waybill
      description: >-
        Atomically proves the quote option is accepted and unexpired, links it to the order,
        and creates or replays exactly one draft waybill identity.
      x-feature-ids: [RATE-07, ORD-05]
      x-permission: order.quote.link
      x-audit-event: order.accepted-quote.linked
      parameters:
        - $ref: '#/components/parameters/OrderId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkAcceptedQuoteToOrderRequest'
      responses:
        '200':
          description: Existing idempotent link replayed.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedQuoteOrderLinkResponse'
        '201':
          description: Quote linked and draft waybill created.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedQuoteOrderLinkResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '422':
          $ref: '#/components/responses/StateTransitionFailed'
...
  /issues:
    post:
      tags: [Issues]
      operationId: createIssue
      summary: Create an operational issue
      x-feature-id: CS-01
      x-permission: issue.manage
      x-audit-event: issue.created
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIssueRequest'
      responses:
        '201':
          description: Issue created independently of waybill main state.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueResponse'
        '422':
          $ref: '#/components/responses/ValidationFailed'
...
  /payments/statement-orders:
    post:
      tags: [Payments]
      operationId: createStatementPaymentOrder
      summary: Create a statement payment order
      x-feature-id: PAY-02
      x-permission: payment.create
      x-audit-event: payment.statement-order.created
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStatementPaymentOrderRequest'
      responses:
        '201':
          description: Payment intent created against frozen statement version.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentOrderResponse'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
...
  /imports:
    get:
      tags: [Imports]
      operationId: listImports
      summary: List import jobs using a stable cursor
      x-feature-id: ORD-03
      x-permission: order.import
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Stable import-job page ordered by createdAt and id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportJobListResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Imports]
      operationId: createImportJob
      summary: Create a durable import job
      x-feature-id: ORD-03
      x-permission: order.import
      x-audit-event: import.created
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateImportJobRequest'
      responses:
        '202':
          description: Import upload accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportJobResponse'
        '422':
          $ref: '#/components/responses/ValidationFailed'
...
  /customers/{customerId}/addresses:upsert:
    parameters:
      - name: customerId
        in: path
        required: true
        schema: {$ref: '#/components/schemas/Ulid'}
    post:
      tags: [Customers]
      operationId: upsertCustomerAddress
      summary: Create or update a customer address in data scope
      description: >-
        The path customerId is authoritative. CREATE forbids If-Match; UPDATE requires a
        strong If-Match and an address id already owned by that path customer.
      x-feature-id: MDM-03
      x-permission: customer.write
      x-audit-event: customer.address.upserted
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/OptionalIfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertCustomerAddressRequest'
      responses:
        '200':
          $ref: '#/components/responses/CommandSucceeded'
        '403':
          $ref: '#/components/responses/Forbidden'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '422':
          $ref: '#/components/responses/StateTransitionFailed'
...
  /documents/exports:
    post:
      tags: [Documents]
      operationId: createExportJob
      summary: Create an audited asynchronous export job
      x-feature-id: DOC-02
      x-permission: document.export
      x-audit-event: document.export.created
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportJob'
      responses:
        '200':
          $ref: '#/components/responses/CommandSucceeded'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/StaleVersion'
        '422':
          $ref: '#/components/responses/StateTransitionFailed'
...
  /portal/api-access-requests:
    post:
      tags: [Portals]
      operationId: createApiAccessRequest
      summary: Request scoped API access for the customer account
      x-feature-id: CUS-03
      x-permission: customer.settings.write
      x-audit-event: portal.api-access-request.created
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiAccessRequest'
      responses:
        '200':
          $ref: '#/components/responses/CommandSucceeded'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/StaleVersion'
        '422':
          $ref: '#/components/responses/StateTransitionFailed'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: zhili_access
  headers:
    ETag:
      description: Quoted aggregate version for If-Match.
      schema:
        type: string
        examples: ['"7"']
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Stable key for the same intended command, retained for at least 24 hours.
      schema:
        type: string
        minLength: 16
        maxLength: 128
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: ETag returned by the latest resource representation.
      schema:
        type: string
        pattern: '^"[1-9][0-9]*"$'
...
    Cursor:
      name: cursor
      in: query
      required: false
      schema: {type: string}
    Limit:
      name: limit
      in: query
      required: false
      schema: {type: integer, minimum: 1, maximum: 200, default: 50}
    OrderId:
      name: orderId
      in: path
      required: true
      schema: {$ref: '#/components/schemas/Ulid'}
...
    QuoteId:
      name: quoteId
      in: path
      required: true
      schema: {$ref: '#/components/schemas/Ulid'}
...
  responses:
...
    Forbidden:
      description: Action, data scope or field policy denied.
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
          example:
            code: PERMISSION_DENIED
            message: Missing action permission.
            details: [{field: permission, reason: warehouse.receive}]
            remediation: Request the listed permission from a tenant administrator.
            requestId: req_01JY8Z8F
...
    PreconditionFailed:
      description: >-
        Missing, malformed or stale strong If-Match precondition. The error code is
        PRECONDITION_REQUIRED for a missing required header and STALE_VERSION for a mismatch.
      content:
        application/problem+json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorEnvelope'
              - type: object
                properties:
                  code: {$ref: '#/components/schemas/PreconditionProblemCode'}
          example:
            code: STALE_VERSION
            message: Data changed after this screen was loaded.
            details: [{field: If-Match, reason: 'expected "7", current "8"'}]
            remediation: Refresh the resource, compare changes, then retry with its latest ETag.
            requestId: req_01JY8Z8F
    ValidationFailed:
      description: Field, restriction or compatibility validation failed.
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...
    Expired:
      description: Quote, impersonation or command intent expired.
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...
  schemas:
...
    Ulid:
      type: string
      pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
      examples: [01JY8Z8F6ME4F0Y9QH2X6D4R7A]
...
    CreateCustomerRequest:
      type: object
      additionalProperties: false
      required: [name, customerCode, settlementCurrency]
      properties:
        name: {type: string}
        customerCode: {type: string}
        settlementCurrency: {$ref: '#/components/schemas/CurrencyCode'}
        creditLimit: {$ref: '#/components/schemas/Money'}
...
    CustomerResponse:
      type: object
      required: [data, meta]
      properties:
        data: {$ref: '#/components/schemas/Customer'}
        meta: {$ref: '#/components/schemas/Meta'}
    CreateOrderDraftRequest:
      type: object
      additionalProperties: false
      required: [orderType, customerId, origin, destination, packages]
      properties:
        orderType: {type: string, enum: [STANDARD, FBA]}
        customerId: {$ref: '#/components/schemas/Ulid'}
        origin: {$ref: '#/components/schemas/Address'}
        destination: {$ref: '#/components/schemas/Address'}
        packages:
          type: array
          minItems: 1
          items: {$ref: '#/components/schemas/PackageInput'}
...
    OrderResponse:
      type: object
      required: [data, meta]
      properties:
        data: {$ref: '#/components/schemas/Order'}
        meta: {$ref: '#/components/schemas/Meta'}
...
    CreateQuoteRequest:
      type: object
      additionalProperties: false
      required: [customerId, origin, destination, packages, quoteDate, currency]
      properties:
        customerId: {$ref: '#/components/schemas/Ulid'}
        origin: {$ref: '#/components/schemas/Address'}
        destination: {$ref: '#/components/schemas/Address'}
        packages:
          type: array
          minItems: 1
          items: {$ref: '#/components/schemas/PackageInput'}
        quoteDate: {type: string, format: date}
        currency: {$ref: '#/components/schemas/CurrencyCode'}
...
    QuoteResponse:
      type: object
      required: [data, meta]
      properties:
        data: {$ref: '#/components/schemas/Quote'}
        meta: {$ref: '#/components/schemas/Meta'}
...
    AcceptQuoteRequest:
      type: object
      additionalProperties: false
      required: [optionId, reason]
      properties:
        optionId: {$ref: '#/components/schemas/Ulid'}
        reason: {type: string, minLength: 2}
...
    QuoteListResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data: {type: array, items: {$ref: '#/components/schemas/Quote'}}
        meta: {$ref: '#/components/schemas/Meta'}
    OrderListResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data: {type: array, items: {$ref: '#/components/schemas/Order'}}
        meta: {$ref: '#/components/schemas/Meta'}
...